refactor(#1195): Separate out IsinGenerator into two parts, the second one dealing with when the system is trying to generate things that aren't ISINs
