+++ generator/src/main/java/com/scottlogic/deg/generator/generation/string/IsinStringGenerator.java
@@ -18,15 +18,11 @@ package com.scottlogic.deg.generator.generation.string;
 
 import com.scottlogic.deg.generator.utils.*;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-// There's a HUGE amount of copy-paste in this class. We should address that before we make any significant changes to it
 public class IsinStringGenerator implements StringGenerator {
     public static final int ISIN_LENGTH = 12;
     private static final String GENERIC_NSIN_REGEX = "[A-Z0-9]{9}";
@@ -35,30 +31,22 @@ public class IsinStringGenerator implements StringGenerator {
     // range of output values when combining with other string generators.
     private RegexStringGenerator isinRegexGenerator;
 
-    private final boolean isNegated;
-
     public IsinStringGenerator() {
-        this(getRegexGeneratorForAllLegalIsinFormats(), false);
+        this(getRegexGeneratorForAllLegalIsinFormats());
     }
 
-    private IsinStringGenerator(RegexStringGenerator regexGenerator, boolean isNegated) {
-        this.isNegated = isNegated;
+    IsinStringGenerator(RegexStringGenerator regexGenerator) {
         isinRegexGenerator = regexGenerator;
     }
 
     @Override
     public StringGenerator intersect(StringGenerator stringGenerator) {
         if (stringGenerator instanceof IsinStringGenerator) {
-            if (isNegated == ((IsinStringGenerator)stringGenerator).isNegated) {
-                RegexStringGenerator otherRegexGenerator =
-                    ((IsinStringGenerator) stringGenerator).isinRegexGenerator;
-                return new IsinStringGenerator(
-                    isNegated
-                        ? isinRegexGenerator.union(otherRegexGenerator)
-                        : (RegexStringGenerator)isinRegexGenerator.intersect(otherRegexGenerator),
-                    isNegated
-                );
-            }
+            return
+                new IsinStringGenerator((RegexStringGenerator)isinRegexGenerator
+                .intersect(((IsinStringGenerator) stringGenerator).isinRegexGenerator));
+        }
+        if (stringGenerator instanceof NotIsinGenerator) {
             return new NoStringsStringGenerator(
                 RegexStringGenerator.intersectRepresentation(stringGenerator.toString(), "<ISIN>")
             );
@@ -80,7 +68,7 @@ public class IsinStringGenerator implements StringGenerator {
 
     private StringGenerator intersect(RegexStringGenerator other) {
         StringGenerator intersection =
-            other.intersect(isNegated ? isinRegexGenerator.complement() : isinRegexGenerator);
+            other.intersect(isinRegexGenerator);
         if ((intersection.isFinite() && intersection.getValueCount() == 0) ||
                 !(intersection instanceof RegexStringGenerator)) {
             return new NoStringsStringGenerator(
@@ -90,23 +78,17 @@ public class IsinStringGenerator implements StringGenerator {
                 )
             );
         }
-        if (!isNegated) {
-            return new IsinStringGenerator((RegexStringGenerator)intersection, false);
-        }
-        return new IsinStringGenerator(
-            isinRegexGenerator.union((RegexStringGenerator)other.complement()),
-            true
-        );
+            return new IsinStringGenerator((RegexStringGenerator)intersection);
     }
 
     @Override
     public StringGenerator complement() {
-        return new IsinStringGenerator(isinRegexGenerator, !isNegated);
+        return new NotIsinGenerator(isinRegexGenerator);
     }
 
     @Override
     public boolean isFinite() {
-        return !isNegated;
+        return true;
     }
 
     @Override
@@ -116,19 +98,11 @@ public class IsinStringGenerator implements StringGenerator {
 
     @Override
     public boolean match(String subject) {
-        boolean matches = FinancialCodeUtils.isValidIsin(subject);
-        return matches != isNegated;
+        return FinancialCodeUtils.isValidIsin(subject);
     }
 
     @Override
     public Iterable<String> generateInterestingValues() {
-        if (isNegated) {
-            return new ConcatenatingIterable<>(
-                Arrays.asList(
-                    generateInterestingInvalidCountryStrings(),
-                    generateInterestingCountriesWithInvalidNsins(),
-                    generateInterestingInvalidCheckDigitIsins()));
-        }
         final List<Iterable<String>> countryCodeIterables = getAllCountryIsinGeneratorsAsStream()
             .limit(2)
             .map(generator -> IterableUtils.wrapIterableWithProjectionAndFilter(
@@ -141,13 +115,6 @@ public class IsinStringGenerator implements StringGenerator {
 
     @Override
     public Iterable<String> generateAllValues() {
-        if (isNegated) {
-            return new ConcatenatingIterable<>(
-                Arrays.asList(
-                    generateAllInvalidCountryStrings(),
-                    generateAllCountriesWithInvalidNsins(),
-                    generateAllInvalidCheckDigitIsins()));
-        }
         final List<Iterable<String>> countryCodeIterables = getAllCountryIsinGeneratorsAsStream()
             .map(generator -> new FilteringIterable<>(
                 generator.generateAllValues(),
@@ -158,14 +125,6 @@ public class IsinStringGenerator implements StringGenerator {
 
     @Override
     public Iterable<String> generateRandomValues(RandomNumberGenerator randomNumberGenerator) {
-        if (isNegated) {
-            return new RandomMergingIterable<>(
-                    Arrays.asList(
-                            generateRandomInvalidCountryStrings(randomNumberGenerator),
-                            generateRandomCountriesWithInvalidNsins(randomNumberGenerator),
-                            generateRandomInvalidCheckDigitIsins(randomNumberGenerator)),
-                    randomNumberGenerator);
-        }
         final List<Iterable<String>> countryCodeIterables = getAllCountryIsinGeneratorsAsStream()
             .map(generator -> IterableUtils.wrapIterableWithProjectionAndFilter(
                 generator.generateRandomValues(randomNumberGenerator),
@@ -179,114 +138,6 @@ public class IsinStringGenerator implements StringGenerator {
     private String replaceCheckDigit(String isin) {
         String isinWithoutCheckDigit = isin.substring(0, isin.length() - 1);
         return isinWithoutCheckDigit + FinancialCodeUtils.calculateIsinCheckDigit(isinWithoutCheckDigit);
-    }
-
-    private Iterable<String> generateInterestingInvalidCountryStrings() {
-        final String invalidCountryCodeRegex = FinancialCodeUtils.VALID_COUNTRY_CODES.stream()
-            .limit(2)
-            .collect(Collectors.joining("|", "((?!", ")).*"));
-        return new RegexStringGenerator(invalidCountryCodeRegex, true).generateInterestingValues();
-    }
-
-    private Iterable<String> generateAllInvalidCountryStrings() {
-        final String invalidCountryCodeRegex = FinancialCodeUtils.VALID_COUNTRY_CODES.stream()
-            .collect(Collectors.joining("|", "((?!", ")).*"));
-        return new RegexStringGenerator(invalidCountryCodeRegex, true).generateAllValues();
-    }
-
-    private Iterable<String> generateRandomInvalidCountryStrings(RandomNumberGenerator randomNumberGenerator) {
-        final String invalidCountryCodeRegex = FinancialCodeUtils.VALID_COUNTRY_CODES.stream()
-            .collect(Collectors.joining("|", "((?!", ")).*"));
-        return new RegexStringGenerator(invalidCountryCodeRegex, true).generateRandomValues(randomNumberGenerator);
-    }
-
-    private static Iterable<String> generateInterestingCountriesWithInvalidNsins() {
-        final List<Iterable<String>> countryWithInvalidNsinIterables = FinancialCodeUtils.VALID_COUNTRY_CODES.stream()
-            .limit(2)
-            .map(countryCode -> {
-                final StringGenerator nsinGeneratorForCountry = getNsinGeneratorForCountry(countryCode);
-                final Iterable<String> invalidNsinIterators = nsinGeneratorForCountry.complement().generateInterestingValues();
-                return new ProjectingIterable<>(invalidNsinIterators, invalidNsin -> countryCode + invalidNsin);
-            })
-            .collect(Collectors.toList());
-        return new FilteringIterable<>(new ConcatenatingIterable<>(countryWithInvalidNsinIterables),
-            isin -> !FinancialCodeUtils.isValidIsin(isin));
-    }
-
-    private static Iterable<String> generateAllCountriesWithInvalidNsins() {
-        final List<Iterable<String>> countryWithInvalidNsinIterables = FinancialCodeUtils.VALID_COUNTRY_CODES.stream()
-                .map(countryCode -> {
-                    final StringGenerator nsinGeneratorForCountry = getNsinGeneratorForCountry(countryCode);
-                    final Iterable<String> invalidNsinIterators = nsinGeneratorForCountry.complement().generateAllValues();
-                    return new ProjectingIterable<>(invalidNsinIterators, invalidNsin -> countryCode + invalidNsin);
-                })
-                .collect(Collectors.toList());
-        return new FilteringIterable<>(new ConcatenatingIterable<>(countryWithInvalidNsinIterables),
-                isin -> !FinancialCodeUtils.isValidIsin(isin));
-    }
-
-    private static Iterable<String> generateRandomCountriesWithInvalidNsins(RandomNumberGenerator randomNumberGenerator) {
-        final List<Iterable<String>> countryWithInvalidNsinIterables = FinancialCodeUtils.VALID_COUNTRY_CODES.stream()
-                .map(countryCode -> {
-                    final StringGenerator nsinGeneratorForCountry = getNsinGeneratorForCountry(countryCode);
-                    final Iterable<String> invalidNsinIterators = nsinGeneratorForCountry.complement().generateRandomValues(randomNumberGenerator);
-                    return new ProjectingIterable<>(invalidNsinIterators, invalidNsin -> countryCode + invalidNsin);
-                })
-                .collect(Collectors.toList());
-        return new FilteringIterable<>(new RandomMergingIterable<>(countryWithInvalidNsinIterables, randomNumberGenerator),
-                isin -> !FinancialCodeUtils.isValidIsin(isin));
-    }
-
-    private Iterable<String> generateInterestingInvalidCheckDigitIsins() {
-        final List<Iterable<String>> countryCodeIterables = getAllCountryIsinGeneratorsAsStream()
-            .limit(2)
-            .map(isinSansCheckDigitGenerator ->
-                new FlatteningIterable<>(
-                    isinSansCheckDigitGenerator.generateInterestingValues(),
-                    isinSansCheckDigit -> {
-                        final char checkDigit = FinancialCodeUtils.calculateIsinCheckDigit(isinSansCheckDigit);
-                        return IntStream.range(0, 10).boxed()
-                            .map(digit -> Character.forDigit(digit, 10))
-                            .filter(digit -> digit != checkDigit)
-                            .map(digit -> isinSansCheckDigit + digit)
-                            .collect(Collectors.toList());
-                    }))
-            .collect(Collectors.toList());
-        return new ConcatenatingIterable<>(countryCodeIterables);
-    }
-
-    private Iterable<String> generateAllInvalidCheckDigitIsins() {
-        final List<Iterable<String>> countryCodeIterables = getAllCountryIsinGeneratorsAsStream()
-                .map(isinSansCheckDigitGenerator ->
-                        new FlatteningIterable<>(
-                                isinSansCheckDigitGenerator.generateAllValues(),
-                                isinSansCheckDigit -> {
-                                    final char checkDigit = FinancialCodeUtils.calculateIsinCheckDigit(isinSansCheckDigit);
-                                    return IntStream.range(0, 10).boxed()
-                                            .map(digit -> Character.forDigit(digit, 10))
-                                            .filter(digit -> digit != checkDigit)
-                                            .map(digit -> isinSansCheckDigit + digit)
-                                            .collect(Collectors.toList());
-                                }))
-                .collect(Collectors.toList());
-        return new ConcatenatingIterable<>(countryCodeIterables);
-    }
-
-    private Iterable<String> generateRandomInvalidCheckDigitIsins(RandomNumberGenerator randomNumberGenerator) {
-        final List<Iterable<String>> countryCodeIterables = getAllCountryIsinGeneratorsAsStream()
-                .map(isinSansCheckDigitGenerator ->
-                        new FlatteningIterable<>(
-                                isinSansCheckDigitGenerator.generateRandomValues(randomNumberGenerator),
-                                isinSansCheckDigit -> {
-                                    final char checkDigit = FinancialCodeUtils.calculateIsinCheckDigit(isinSansCheckDigit);
-                                    return IntStream.range(0, 10).boxed()
-                                            .map(digit -> Character.forDigit(digit, 10))
-                                            .filter(digit -> digit != checkDigit)
-                                            .map(digit -> isinSansCheckDigit + digit)
-                                            .collect(Collectors.toList());
-                                }))
-                .collect(Collectors.toList());
-        return new RandomMergingIterable<>(countryCodeIterables, randomNumberGenerator);
     }
 
     private Stream<StringGenerator> getAllCountryIsinGeneratorsAsStream() {
@@ -334,27 +185,17 @@ public class IsinStringGenerator implements StringGenerator {
         return getIsinRegexRepresentationForCountrySansCheckDigit(countryCode) + "[0-9]";
     }
 
-    private static StringGenerator getNsinGeneratorForCountry(String countryCode) {
-        if (countryCode.equals("GB")) {
-            return new SedolStringGenerator("00");
-        }
-        if (countryCode.equals("US")) {
-            return new CusipStringGenerator();
-        }
-        return new RegexStringGenerator(GENERIC_NSIN_REGEX, true);
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
         IsinStringGenerator that = (IsinStringGenerator) o;
-        return isNegated == that.isNegated;
+        return Objects.equals(this.isinRegexGenerator, that.isinRegexGenerator);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(isNegated, getClass());
+        return Objects.hash(isinRegexGenerator);
     }
 }
